+++ src/Models/Concerns/HasShippingAddresses.php
@@ -19,6 +19,10 @@ trait HasShippingAddresses
         }
 
         static::deleted(function (Model $deletedModel) {
+            if (method_exists($deletedModel, 'isForceDeleting') && ! $deletedModel->isForceDeleting()) {
+                return;
+            }
+
             $deletedModel->shippingAddresses()->delete();
         });
     }
